.claude/worktrees/goofy-dewdney/requirements.txt
requests
pandas
supabase
pykrx
urllib3
groq
opendartreader
python-dotenv
beautifulsoup4
openpyxl
fastapi
uvicorn[standard]
pydantic